Move sqlalchemy_fauna to root of tipping module
Trying to move this outside of tipping and getting all the
imports to play nice is more trouble than it's worth right now.
I may look into installing it as a package in editable mode when
I'm closer to to making it a standalone module.
