--- tipping/src/tipping/db/__init__.py
@@ -1,1 +1,0 @@
-"""Manages configuration for and connections to databases."""
