--- tipping/src/tipping/db/sqlalchemy_fauna/__init__.py
@@ -1,6 +1,0 @@
-"""Creates a Fauna Dialect for SQLAlchemy."""
-
-from tipping.db.sqlalchemy_fauna.dbapi import connect
-from tipping.db.sqlalchemy_fauna.exceptions import Error, NotSupportedError
-
-paramstyle = "pyformat"
