+++ tipping/src/tipping/sqlalchemy_fauna/__init__.py
@@ -1,0 +1,6 @@
+"""Creates a Fauna Dialect for SQLAlchemy."""
+
+from tipping.sqlalchemy_fauna.dbapi import connect
+from tipping.sqlalchemy_fauna.exceptions import Error, NotSupportedError
+
+paramstyle = "pyformat"
