+++ tipping/src/tipping/sqlalchemy_fauna/exceptions.py
